Update .gitignore and minishell.c
