+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: miheider <miheider@student.42.fr>          +#+  +:+       +#+         #
+#    By: caigner <caigner@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/03/10 20:04:47 by miheider          #+#    #+#              #
-#    Updated: 2024/03/23 14:59:07 by miheider         ###   ########.fr        #
+#    Updated: 2024/03/23 19:18:10 by caigner          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,8 +31,6 @@ SOURCES	=	src/builtins/ft_echo.c	\
 			src/minishell.c \
 			src/init_env.c \
 			src/lexer.c \
-			src/lexer_array.c \
-			src/lexer_prep_input.c \
 			src/lexer_utils.c \
 			src/parsing.c \
 			src/expansion.c \
