+++ src/minishell.c
@@ -6,7 +6,7 @@
 /*   By: caigner <caigner@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/28 20:17:44 by chris             #+#    #+#             */
-/*   Updated: 2024/03/23 16:34:35 by caigner          ###   ########.fr       */
+/*   Updated: 2024/03/23 19:14:35 by caigner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ int	ft_loop(t_common *c)
 	while (1)
 	{
 		init_loop_data(c);
-		interactive();
+//		interactive();
 		c->raw_prompt = prompt(c);
 //		ft_putstr_fd(c->raw_prompt, 1);
 //		if (!c->raw_prompt[0])
